SBERDOMA-62 refactor: extract checkBillingIntegrationAccessRight
